# Split: Reduce memory allocation