--- a/reedsolomon.go
+++ b/reedsolomon.go
@@ -818,15 +818,26 @@ func (r reedSolomon) Split(data []byte) ([][]byte, error) {
 	}
 
 	// Only allocate memory if necessary
+	var padding []byte
 	if len(data) < (r.Shards * perShard) {
-		data = append(data, make([]byte, r.Shards*perShard-len(data))...)
+		// calculate maximum number of full shards in `data` slice
+		fullShards := len(data) / perShard
+		padding = make([]byte, r.Shards*perShard-perShard*fullShards)
+		copy(padding, data[perShard*fullShards:])
+		data = data[0 : perShard*fullShards]
 	}
 
 	// Split into equal-length shards.
 	dst := make([][]byte, r.Shards)
-	for i := range dst {
+	i := 0
+	for ; i < len(dst) && len(data) >= perShard; i++ {
 		dst[i] = data[:perShard]
 		data = data[perShard:]
+	}
+
+	for j := 0; i+j < len(dst); j++ {
+		dst[i+j] = padding[:perShard]
+		padding = padding[perShard:]
 	}
 
 	return dst, nil