The following part of code has 2 memory allocation:
```go
if len(data) < (r.Shards * perShard) {
    // Pad data to r.Shards*perShard.
    padding := make([]byte, (r.Shards*perShard)-len(data))
    data = append(data, padding...)
}
```

1. `make([]byte, (r.Shards*perShard)-len(data))`
2. `append` function allocates new slice with `capacity = (len(data) + len(padding) + 1)*2` ( More details https://blog.golang.org/go-slices-usage-and-internals )

The following line of code is fixed the problem:
```go
data = append(data, make([]byte, r.Shards*perShard - len(data))...)
```
Because compiler specially optimize this case: https://github.com/golang/go/issues/21266


Old code benchmark:
```go
BenchmarkSplit10x4x160M-4   	      30	  54329722 ns/op	329253216 B/op	       3 allocs/op
BenchmarkSplit5x2x5M-4      	    1000	   1616534 ns/op	10289328 B/op	       3 allocs/op
BenchmarkSplit10x2x1M-4     	    5000	    212577 ns/op	 1524000 B/op	       3 allocs/op
BenchmarkSplit10x4x10M-4    	     500	   3085607 ns/op	20578657 B/op	       3 allocs/op
BenchmarkSplit50x20x50M-4   	     100	  16650049 ns/op	102893316 B/op	       3 allocs/op
BenchmarkSplit17x3x272M-4   	      20	  71213362 ns/op	406847972 B/op	       3 allocs/op
```

New code benchmark:
```
BenchmarkSplit10x4x160M-4   	      20	  50570930 ns/op	262144353 B/op	       2 allocs/op
BenchmarkSplit5x2x5M-4      	    2000	   1212313 ns/op	 8192176 B/op	       2 allocs/op
BenchmarkSplit10x2x1M-4     	   10000	    182420 ns/op	 1311008 B/op	       2 allocs/op
BenchmarkSplit10x4x10M-4    	     500	   2255361 ns/op	16384352 B/op	       2 allocs/op
BenchmarkSplit50x20x50M-4   	     100	  13558897 ns/op	81921792 B/op	       2 allocs/op
BenchmarkSplit17x3x272M-4   	      30	  64837941 ns/op	356516323 B/op	       2 allocs/op
```

Compare new and old
```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSplit10x4x160M-4     54329722      50570930      -6.92%
BenchmarkSplit5x2x5M-4        1616534       1212313       -25.01%
BenchmarkSplit10x2x1M-4       212577        182420        -14.19%
BenchmarkSplit10x4x10M-4      3085607       2255361       -26.91%
BenchmarkSplit50x20x50M-4     16650049      13558897      -18.57%
BenchmarkSplit17x3x272M-4     71213362      64837941      -8.95%

benchmark                     old allocs     new allocs     delta
BenchmarkSplit10x4x160M-4     3              2              -33.33%
BenchmarkSplit5x2x5M-4        3              2              -33.33%
BenchmarkSplit10x2x1M-4       3              2              -33.33%
BenchmarkSplit10x4x10M-4      3              2              -33.33%
BenchmarkSplit50x20x50M-4     3              2              -33.33%
BenchmarkSplit17x3x272M-4     3              2              -33.33%

benchmark                     old bytes     new bytes     delta
BenchmarkSplit10x4x160M-4     329253216     262144353     -20.38%
BenchmarkSplit5x2x5M-4        10289328      8192176       -20.38%
BenchmarkSplit10x2x1M-4       1524000       1311008       -13.98%
BenchmarkSplit10x4x10M-4      20578657      16384352      -20.38%
BenchmarkSplit50x20x50M-4     102893316     81921792      -20.38%
BenchmarkSplit17x3x272M-4     406847972     356516323     -12.37%
```