@klauspost 2019-06-21
Thanks! I never spent much time on these, since people tend to do their own :)

But it seems great! I will just do a final check and merge in the weekend!