@dssysolyatin 2019-06-21
@klauspost I tried to further reduce memory allocation. I have done it:

```
benchmark                     old ns/op     new ns/op     delta
BenchmarkSplit10x4x160M-4     54329722      8564593       -84.24%
BenchmarkSplit5x2x5M-4        1616534       345720        -78.61%
BenchmarkSplit10x2x1M-4       212577        61885         -70.89%
BenchmarkSplit10x4x10M-4      3085607       601724        -80.50%
BenchmarkSplit50x20x50M-4     16650049      2703954       -83.76%
BenchmarkSplit17x3x272M-4     71213362      6942278       -90.25%

benchmark                     old allocs     new allocs     delta
BenchmarkSplit10x4x160M-4     3              2              -33.33%
BenchmarkSplit5x2x5M-4        3              2              -33.33%
BenchmarkSplit10x2x1M-4       3              2              -33.33%
BenchmarkSplit10x4x10M-4      3              2              -33.33%
BenchmarkSplit50x20x50M-4     3              2              -33.33%
BenchmarkSplit17x3x272M-4     3              2              -33.33%

benchmark                     old bytes     new bytes     delta
BenchmarkSplit10x4x160M-4     329253216     67109216      -79.62%
BenchmarkSplit5x2x5M-4        10289328      2097328       -79.62%
BenchmarkSplit10x2x1M-4       1524000       319776        -79.02%
BenchmarkSplit10x4x10M-4      20578657      4194656       -79.62%
BenchmarkSplit50x20x50M-4     102893316     20973312      -79.62%
BenchmarkSplit17x3x272M-4     406847972     50332128      -87.63%
```